DABOM-487 refactor: usage 서비스 warmup 연동 수정 및 테스트 정리
