+++ src/main/java/com/project/domain/usage/service/UsagePersistServiceImpl.java
@@ -15,7 +15,7 @@ import com.project.domain.usage.service.helper.CustomerQuotaWriter;
 import com.project.domain.usage.service.helper.FamilyQuotaWriter;
 import com.project.domain.usage.service.helper.UsagePersistEventValidator;
 import com.project.domain.usage.service.helper.UsageRecordWriter;
-import com.project.global.common.TimeConstants;
+import com.project.global.config.TimeConfig;
 import com.project.global.util.LogSanitizer;
 
 import lombok.RequiredArgsConstructor;
@@ -101,14 +101,14 @@ public class UsagePersistServiceImpl implements UsagePersistService {
 
     // 이벤트 시각을 정산 월로 변환하고 이상 값이면 현재 월로 보정한다.
     private LocalDate resolveCurrentMonth(String eventTime) {
-        LocalDate currentMonth = LocalDate.now(TimeConstants.ASIA_SEOUL).withDayOfMonth(1);
+        LocalDate currentMonth = LocalDate.now(TimeConfig.ASIA_SEOUL).withDayOfMonth(1);
         if (eventTime == null || eventTime.isBlank()) {
             return currentMonth;
         }
         try {
             LocalDate parsedMonth =
                     LocalDateTime.parse(eventTime)
-                            .atZone(TimeConstants.ASIA_SEOUL)
+                            .atZone(TimeConfig.ASIA_SEOUL)
                             .toLocalDate()
                             .withDayOfMonth(1);
