+++ src/main/java/com/project/domain/usage/service/helper/UsageRedisWarmupHelper.java
@@ -20,7 +20,7 @@ import com.project.domain.family.repository.FamilyQuotaRepository;
 import com.project.domain.family.repository.FamilyRepository;
 import com.project.domain.usage.infra.cache.dto.FamilyInfoRedisHash;
 import com.project.domain.usage.service.dto.FamilyInfo;
-import com.project.global.common.TimeConstants;
+import com.project.global.config.TimeConfig;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -159,7 +159,7 @@ public class UsageRedisWarmupHelper {
 
             long usedBytes = (quota == null) ? 0L : Math.max(0L, quota.getMonthlyUsedBytes());
             long nextMonthStartEpochSecond =
-                    eventMonth.plusMonths(1).atStartOfDay(TimeConstants.ASIA_SEOUL).toEpochSecond();
+                    eventMonth.plusMonths(1).atStartOfDay(TimeConfig.ASIA_SEOUL).toEpochSecond();
 
             // 키가 없어도 월초 첫 트래픽을 안전하게 처리하도록 0으로 시드하고 만료를 설정함
             Boolean written =
